Added logic to update balance

+++ src/main/kotlin/com/personal/toolchest/service/impl/TransactionServiceImpl.kt
@@ -14,10 +14,17 @@ class TransactionServiceImpl: TransactionService {
     @Autowired
     private lateinit var transactionRepository: TransactionRepository
 
+    @Autowired
+    private lateinit var transactionTypeServiceImpl: TransactionTypeServiceImpl
+
     override fun createTransaction(transaction: TransactionDTO) {
         val transactionEntity = Transaction(
-            name = transaction.name, amount = transaction.amount, transactionType = transaction.transactionTypeEnum)
+            name = transaction.name, amount = transaction.amount, transactionType = transaction.transactionType)
         transactionRepository.save(transactionEntity)
+        val currentBalance =
+            transactionTypeServiceImpl.getCurrentBalance(transaction.transactionType)
+        val newBalance = currentBalance.subtract(transaction.amount)
+        transactionTypeServiceImpl.updateBalance(transaction.transactionType, newBalance)
     }
 
     override fun getTransactions(deleted: Boolean): List<TransactionDTO> {
@@ -27,10 +34,17 @@ class TransactionServiceImpl: TransactionService {
     }
 
     override fun deleteTransaction(transactionId: Int) {
-        transactionRepository.deleteById(transactionId)
+        val transaction = transactionRepository.findById(transactionId);
+        if (transaction.isPresent) {
+            transactionRepository.deleteById(transactionId)
+            val transactionType = transaction.get().transactionType
+            val currentBalance =
+                transactionTypeServiceImpl.getCurrentBalance(transactionType)
+            val newBalance = currentBalance.add(transaction.get().amount)
+            transactionTypeServiceImpl.updateBalance(transactionType, newBalance)
+        }
     }
 
     override fun getMonthlySpent() {
-
     }
 }
